initial commit
In this scenario, a coupon recommendation dataset is taken into consideration containing   several features(including customer/drivers’ family status, his interests, present occupation or earnings, etc) to classify that whatever the coupon is recommended to the person(who is driver) will be accepted by him or not. For this kind of classification task, features play a vital role as whatever the outcome is estimated  barely depends over the circumstances or features that support the outcome, the most. 
Out of the proposed algorithms i.e, Decision Tree classifier, XGBoost classifier, and Artificial Neural Networks, conclusion is given that which algorithm comes out of the box to classify the same, excellently.